refact: retrieve part of rendering Focus units to isolated dir in Main.
